integrate datastream and evaluator agents into single file
