+++ agentMET4FOF_ml_extension/tutorials/ml_experiment_complex.py
@@ -8,21 +8,17 @@ hyperparameter range are made too large.
 
 
 from agentMET4FOF.agents import AgentNetwork, MonitorAgent
+
 from agentMET4FOF_ml_extension.Dashboard_ml_exp import Dashboard_ML_Experiment
-from agentMET4FOF_ml_extension.datastream import *
-from agentMET4FOF_ml_extension.evaluator import *
-from agentMET4FOF_ml_extension.ML_Experiment import *
-from agentMET4FOF_ml_extension.agents import AgentPipeline
+from agentMET4FOF_ml_extension.ML_Experiment import ML_Experiment
+from agentMET4FOF_ml_extension.agents import AgentPipeline, ML_DataStreamAgent, ML_EvaluatorAgent
 
 from sklearn import datasets
 from sklearn.metrics import f1_score
 from sklearn.decomposition import PCA
 from sklearn.preprocessing import StandardScaler, RobustScaler, MinMaxScaler, MaxAbsScaler,PowerTransformer
-
-
 from sklearn.gaussian_process import GaussianProcessClassifier
 from sklearn.gaussian_process.kernels import RBF
-from agentMET4FOF_ml_extension.ml_uncertainty.evaluate_pred_unc import *
 
 def main():
     #initialise with dashboard ml experiments
