+++ agentMET4FOF_ml_extension/tutorials/with_bnn/ml_experiment_simple.py
@@ -11,12 +11,12 @@ consisting of pipelines and subsets called chains, can be viewed and compared.
 
 
 from agentMET4FOF.agents import AgentNetwork, MonitorAgent
-from agentMET4FOF_ml_extension.datastream import *
-from agentMET4FOF_ml_extension.evaluator import *
-from agentMET4FOF_ml_extension.ML_Experiment import *
+
+from agentMET4FOF_ml_extension.Dashboard_ml_exp import Dashboard_ML_Experiment
+from agentMET4FOF_ml_extension.ML_Experiment import ML_Experiment
+from agentMET4FOF_ml_extension.agents import AgentPipeline, ML_DataStreamAgent, ML_EvaluatorAgent
 from agentMET4FOF_ml_extension.ml_uncertainty.bnn import BNN_Model
 from agentMET4FOF_ml_extension.ml_uncertainty.evaluate_pred_unc import *
-from agentMET4FOF_ml_extension.agents import AgentPipeline
 
 from sklearn import datasets
 from sklearn.preprocessing import StandardScaler
